Reduce flash usage on arduino_main sketch
- Guard Serial.begin() behind HAS_CIRCUIT_PLAYGROUND so the UART
  library is dropped by the linker when the CP is not connected
- Simplify findNextMountain: remove dead second loop that could
  never return a different result than the first sequential check
- Replace long arithmetic with int in calculateServoAngle; all
  values fit in 16-bit so 32-bit ops were unnecessary on AVR
- Fix accidental active #define FAST_TEST_MODE in progress.h that
  was capping every mountain at 3 sessions for all users

+++ firmware/arduino_main/servo_control.h
@@ -53,8 +53,8 @@ inline int calculateServoAngle(uint8_t sessions, uint8_t totalSessions) {
   if (totalSessions == 0) return SERVO_MIN_ANGLE;
   if (sessions >= totalSessions) return SERVO_MAX_ANGLE;
 
-  long range = SERVO_MAX_ANGLE - SERVO_MIN_ANGLE; // 150
-  return SERVO_MIN_ANGLE + (int)((range * (long)sessions) / (long)totalSessions);
+  int range = SERVO_MAX_ANGLE - SERVO_MIN_ANGLE; // 150
+  return SERVO_MIN_ANGLE + (range * (int)sessions) / (int)totalSessions;
 }
 
 /**
